+++ src/metadata.rs
@@ -332,10 +332,92 @@ pub mod testing {
 mod tests {
   use super::*;
   use serde_json;
+  use std::fs::File;
+  use std::io::Write;
+
+  fn basic_toml() -> &'static str {
+    "
+[package]
+name = \"test\"
+version = \"0.0.1\"
+
+[lib]
+path = \"not_a_file.rs\"
+    "
+  }
+
+  fn basic_lock() -> &'static str {
+    "
+[[package]]
+name = \"test\"
+version = \"0.0.1\"
+dependencies = [
+]
+    "
+  }
 
   #[test]
   fn test_metadata_deserializes_correctly() {
     let metadata_file_contents = include_str!("../test_fixtures/metadata.txt");
-    let metadata = serde_json::from_str::<Metadata>(metadata_file_contents).unwrap();
+    serde_json::from_str::<Metadata>(metadata_file_contents).unwrap();
+  }
+
+  #[test]
+  fn test_cargo_subcommand_metadata_fetcher_works_without_lock() {
+    let dir = TempDir::new("test_cargo_raze_metadata_dir").unwrap();
+    let toml_path = dir.path().join("Cargo.toml");
+    let mut toml = File::create(&toml_path).unwrap();
+    toml.write_all(basic_toml().as_bytes()).unwrap();
+    let files = CargoWorkspaceFiles {
+      toml_path: toml_path,
+      lock_path_opt: None,
+    };
+
+    let mut fetcher = CargoSubcommandMetadataFetcher;
+
+    fetcher.fetch_metadata(files).unwrap();
+  }
+
+  #[test]
+  fn test_cargo_subcommand_metadata_fetcher_works_with_lock() {
+    let dir = TempDir::new("test_cargo_raze_metadata_dir").unwrap();
+    let toml_path = {
+      let path = dir.path().join("Cargo.toml");
+      let mut toml = File::create(&path).unwrap();
+      toml.write_all(basic_toml().as_bytes()).unwrap();
+      path
+    };
+    let lock_path = {
+      let path = dir.path().join("Cargo.lock");
+      let mut lock = File::create(&path).unwrap();
+      lock.write_all(basic_lock().as_bytes()).unwrap();
+      path
+    };
+    let files = CargoWorkspaceFiles {
+      toml_path: toml_path,
+      lock_path_opt: Some(lock_path),
+    };
+
+    let mut fetcher = CargoSubcommandMetadataFetcher;
+
+    fetcher.fetch_metadata(files).unwrap();
+  }
+
+  #[test]
+  fn test_cargo_subcommand_metadata_fetcher_handles_bad_files() {
+    let dir = TempDir::new("test_cargo_raze_metadata_dir").unwrap();
+    let toml_path = {
+      let path = dir.path().join("Cargo.toml");
+      let mut toml = File::create(&path).unwrap();
+      toml.write_all(b"hello").unwrap();
+      path
+    };
+    let files = CargoWorkspaceFiles {
+      toml_path: toml_path,
+      lock_path_opt: None,
+    };
+
+    let mut fetcher = CargoSubcommandMetadataFetcher;
+    assert!(fetcher.fetch_metadata(files).is_err());
   }
 }
